+++ .eslintrc.js
@@ -81,11 +81,6 @@ module.exports = {
             name: 'Buffer',
             message: 'Global buffer is not supported on all platforms. Import buffer from `src/utils/buffer`',
           },
-          {
-            name: 'AbortController',
-            message:
-              "Global AbortController is not supported on all platforms. Use `import { AbortController } from 'abort-controller'`",
-          },
         ],
       },
     },