+++ packages/webvh/src/dids/WebVhDidRegistrar.ts
@@ -72,7 +72,7 @@ export class WebVhDidRegistrar implements DidRegistrar {
         verificationMethods: [
           {
             controller: baseDid,
-            type: 'Ed25519VerificationKey2020',
+            type: 'Multikey',
             publicKeyMultibase,
           },
         ],
@@ -151,10 +151,7 @@ export class WebVhDidRegistrar implements DidRegistrar {
         domain,
         updateKeys,
         ...inputDidDocument,
-        verificationMethods: verificationMethods.map((vm) => ({
-          ...vm,
-          publicKeyMultibase: vm.publicKeyMultibase ?? '',
-        })),
+        verificationMethods,
         controller: Array.isArray(controller) ? controller[0] : controller,
         authentication: normalizeMethodArray(authentication),
         assertionMethod: normalizeMethodArray(assertionMethod),